+++ projects/demo/server.ts
@@ -1,9 +1,9 @@
 import { APP_BASE_HREF } from '@angular/common';
+import { renderApplication } from '@angular/platform-server';
 import express from 'express';
 import { fileURLToPath } from 'node:url';
 import { dirname, join, resolve } from 'node:path';
-import { renderModule } from '@angular/platform-server';
-import { ThemeToggleModule } from '@dasdigitalplatform/dls-global-angular/theme-toggle';
+import bootstrap from './src/main.server';
 
 
 
@@ -13,8 +13,6 @@ export function app(): express.Express {
   const serverDistFolder = dirname(fileURLToPath(import.meta.url));
   const browserDistFolder = resolve(serverDistFolder, '../browser');
   const indexHtml = join(serverDistFolder, 'index.server.html');
-
-  // const commonEngine = new CommonEngine();
 
   server.set('view engine', 'html');
   server.set('views', browserDistFolder);
@@ -30,14 +28,13 @@ export function app(): express.Express {
   server.get('*', (req, res, next) => {
     const { protocol, originalUrl, baseUrl, headers } = req;
 
-    // Update the rendering logic
-    renderModule(ThemeToggleModule, {
+    renderApplication(bootstrap, {
       document: indexHtml,
       url: `${protocol}://${headers.host}${originalUrl}`,
-      extraProviders: [{ provide: APP_BASE_HREF, useValue: baseUrl }],
+      platformProviders: [{ provide: APP_BASE_HREF, useValue: baseUrl }],
     })
-      .then((html) => res.send(html))
-      .catch((err) => next(err));
+      .then((html: any) => res.send(html))
+      .catch((err: any) => next(err));
   });
 
   return server;