+++ projects/demo/src/app/breadcrumb/breadcrumb.component.ts
@@ -1,15 +1,14 @@
 import { Component } from '@angular/core';
+import { BreadcrumbComponent } from '../../../../design/dls-global-angular/breadcrumb/src/breadcrumb.component';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatButtonModule } from '@angular/material/button';
-import { BreadcrumbComponent } from '@dasdigitalplatform/dls-global-angular/breadcrumb';
-import { IconRegistryModule } from '@dasdigitalplatform/dls-global-angular/icon-registry';
 
 @Component({
     selector: 'breadcrumb-page',
-    imports: [MatButtonModule, BreadcrumbComponent, IconRegistryModule],
+    imports: [MatButtonModule, BreadcrumbComponent],
     templateUrl: './breadcrumb.component.html',
     styleUrl: './breadcrumb.component.scss'
-})
+}) 
 export class BreadcrumbPageComponent {
   title = 'Breadcrumb';