+++ projects/design/dls-global-angular/sidesheet/src/sidesheet.component.ts
@@ -2,15 +2,15 @@ import { Component, Input } from '@angular/core';
 import { MatIconModule } from '@angular/material/icon';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatButtonModule } from '@angular/material/button';
-import { MatDividerModule } from '@angular/material/divider';
 import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatDividerModule } from '@angular/material/divider';
 import {
   AtmosphereSidesheetVariant,
   AtmosphereSidesheetContentMode,
   AtmosphereSidesheetButtonsArray
 } from './sidesheet.types';
 
-import {IconRegistryModule} from "@dasdigitalplatform/dls-global-angular/icon-registry";
+import { IconRegistryModule } from "@dasdigitalplatform/dls-global-angular/icon-registry";
 
 export enum SidesheetVariant {
   STANDARD = 'standard',