+++ projects/design/dls-global-angular/stories/Angular Components/components/navigationdrawer/navigationDrawer.stories.ts
@@ -1,12 +1,13 @@
-import { Meta, moduleMetadata, StoryObj} from '@storybook/angular';
+import { Meta, moduleMetadata, StoryObj } from '@storybook/angular';
 import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatDividerModule } from '@angular/material/divider';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import {
   NavigationDrawerComponent,
   MockSidebarMenuData,
-  NavDrawerPresentationEnum
 } from '../../../../navigation-drawer';
 import { CommonModule } from '@angular/common';
 import { IconRegistryStorybookModule } from '../../../../icon-registry';
@@ -29,7 +30,7 @@ const mockActivatedRoute = {
 };
 
 const basicTemplate = (type: string): string =>
-  ` 
+  `
     <ba-navigation-drawer
       [fixedOpen]="fixedOpen"
       [minifyOnCollapse]="minifyOnCollapse"
@@ -38,13 +39,13 @@ const basicTemplate = (type: string): string =>
       [navBarBottomComponent2] = "sideNavProfileButton"
       [appName]="'Boeing App'"
       [appLogo]="'boeing'"
-      [opened]="opened"
+      [(opened)]="opened"
       [homeRoute]="homeRoute"
      ></ba-navigation-drawer>
 
     <ng-template #sideNavNotificationsButton let-data="data">
-      <button 
-        mat-icon-button 
+      <button
+        mat-icon-button
         [ngClass]="{
           'mat-mdc-icon-button--toggle-on': toggled === true
         }"
@@ -54,7 +55,33 @@ const basicTemplate = (type: string): string =>
     </ng-template>
 
     <ng-template #sideNavProfileButton let-data="data">
-      <ba-user-profile clickable userFirstName="William"></ba-user-profile>
+      <ba-user-profile clickable [userFirstName]="firstName" userAvatarSource="./assets/card-img-1.png" [matMenuTriggerFor]="menuUserImage"></ba-user-profile>
+    <mat-menu #menuUserImage="matMenu">
+      <button mat-menu-item class="user-info-menu-item">
+        <ba-user-profile [userFirstName]="firstName" userAvatarSource="./assets/card-img-1.png"></ba-user-profile>
+        <div class="user-info">
+          <span class="user-name">{{ firstName }}</span>
+          <span class="additional-info">Additional info</span>
+        </div>
+      </button>
+      <mat-divider></mat-divider>
+      <button mat-menu-item>
+        <span>Profile</span>
+      </button>
+      <button mat-menu-item>
+        <span>Preferences</span>
+      </button>
+      <mat-divider></mat-divider>
+      <button mat-menu-item>
+        <mat-icon svgIcon="icon-question-mark">question-mark</mat-icon>
+        <span>Help and tutorials</span>
+      </button>
+      <mat-divider></mat-divider>
+      <button mat-menu-item>
+        <mat-icon svgIcon="icon-power-settings-new">power-settings-new</mat-icon>
+        <span>Sign out</span>
+      </button>
+    </mat-menu>
     </ng-template>
   `
 
@@ -70,6 +97,8 @@ export default {
         BrowserAnimationsModule,
         MatButtonModule,
         MatIconModule,
+        MatMenuModule,
+        MatDividerModule,
         IconRegistryStorybookModule,
         MatListModule,
         RouterModule,
@@ -104,7 +133,7 @@ export const Default: Story = {
     opened: true,
     minifyOnCollapse: true,
   },
-  render: ({ opened, minifyOnCollapse, homeRoute }) => ({
+  render: ({ opened, minifyOnCollapse, homeRoute }: { opened: boolean; minifyOnCollapse: boolean; homeRoute?: string }) => ({
     props: {
       opened,
       minifyOnCollapse, homeRoute
@@ -123,7 +152,7 @@ export const Standalone: Story = {
     opened: false,
     minifyOnCollapse: true,
   },
-  render: ({ opened, minifyOnCollapse, homeRoute }) => ({
+  render: ({ opened, minifyOnCollapse, homeRoute }: { opened: boolean; minifyOnCollapse: boolean; homeRoute?: string }) => ({
     props: { opened, minifyOnCollapse, homeRoute },
     template: basicTemplate('standalone')
   }),
@@ -140,7 +169,7 @@ export const Modal: Story = {
     minifyOnCollapse: true,
     fixedOpen: false,
   },
-  render: ({ opened, fixedOpen, homeRoute, minifyOnCollapse }) => ({
+  render: ({ opened, fixedOpen, homeRoute, minifyOnCollapse }: { opened: boolean; fixedOpen: boolean; homeRoute?: string; minifyOnCollapse: boolean }) => ({
     props: { opened, fixedOpen, homeRoute, minifyOnCollapse },
     template: basicTemplate('modal')
   }),
@@ -156,7 +185,7 @@ export const FixedOpen: Story = {
     opened: true,
     fixedOpen: true,
   },
-  render: ({ opened, fixedOpen, homeRoute }) => ({
+  render: ({ opened, fixedOpen, homeRoute }: { opened: boolean; fixedOpen: boolean; homeRoute?: string }) => ({
     props: { opened, fixedOpen, homeRoute },
     template: basicTemplate('default')
   }),