+++ projects/design/dls-global-angular/theme-toggle/src/theme-toggle.component.ts
@@ -16,10 +16,11 @@ import { MatIcon } from '@angular/material/icon';
   selector: "ba-theme-toggle",
   imports: [MatIconButton, MatTooltip, MatIcon],
   templateUrl: "theme-toggle.component.html",
-  styleUrls: ["theme-toggle.component.scss"]
+  styleUrls: ["theme-toggle.component.scss"],
+  standalone: true,
 })
 export class ThemeToggleComponent {
-  constructor(private themeToggleService: ThemeToggleService) { }
+  constructor(private themeToggleService: ThemeToggleService) {}
 
   toggle() {
     this.themeToggleService.toggleTheme();