+++ projects/design/dls-global-angular/theme-toggle/src/theme-toggle.service.ts
@@ -79,6 +79,15 @@ export class ThemeToggleService {
   }
 
   /**
+   * Function to get the current theme
+   * @returns Current theme
+   * Exposed publicly
+   */
+  getCurrentTheme(): Theme {
+    return this.currentTheme;
+  }
+
+  /**
    * Function that toggles the theme
    * Exposed publicly
    */