+++ projects/design/dls-global-angular/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.3.0](https://git.web.boeing.com/design/design-system/dls-global-angular/compare/v3.2.2...v3.3.0) (2025-07-21)
+
+
+### Features
+
+* Allow users to set the title in the sidesheet ([587930d](https://git.web.boeing.com/design/design-system/dls-global-angular/commits/587930d863919612fa22f997a0aaf90e45fdd844))
+
+
+
+# [3.3.0](https://git.web.boeing.com/design/design-system/dls-global-angular/compare/v3.2.2...v3.3.0) (2025-07-21)
+
+
+### Features
+
+* Allow users to set the title in the sidesheet ([587930d](https://git.web.boeing.com/design/design-system/dls-global-angular/commits/587930d863919612fa22f997a0aaf90e45fdd844))
+
+
+
+# [3.3.0](https://git.web.boeing.com/design/design-system/dls-global-angular/compare/v3.2.2...v3.3.0) (2025-07-21)
+
+
+### Features
+
+* Allow users to set the title in the sidesheet ([587930d](https://git.web.boeing.com/design/design-system/dls-global-angular/commits/587930d863919612fa22f997a0aaf90e45fdd844))
+
+
+
 ## [3.2.2](https://git.web.boeing.com/design/design-system/dls-global-angular/compare/v3.2.1...v3.2.2) (2025-05-20)
 
 
@@ -10,3 +37,24 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 * **nav-drawer:** enhnenching the quality of navigation drawer. ([d9af466](https://git.web.boeing.com/design/design-system/dls-global-angular/commits/d9af466cd05126d63e8bb00bdc806f046cdf1a4d))
 * **navigation-drawer:** resolve UI inconsistencies and enhacne functionality for angualr 19 ([4e11492](https://git.web.boeing.com/design/design-system/dls-global-angular/commits/4e11492a1ca2385a9a674ad5fd2e68cb53888569))
+
+
+# [2.6.0](https://git.web.boeing.com/design/design-system/dls-global-angular/compare/v2.5.10...v2.6.0) (2025-05-15)
+
+
+### Bug Fixes
+
+* icon alignment ([052ade8](https://git.web.boeing.com/design/design-system/dls-global-angular/commits/052ade81a6cd28d7dbb8b5f59b68d84b4cb3c577))
+
+
+### Features
+
+* add tooltip for navigation rail ([2254cc1](https://git.web.boeing.com/design/design-system/dls-global-angular/commits/2254cc13ac85ed63a4e3aa536ee6360570291299))
+
+
+
+
+
+## [2.5.10](https://git.web.boeing.com/design/design-system/dls-global-angular/compare/v2.5.9...v2.5.10) (2025-05-09)
+
+**Note:** Version bump only for package @dips/dls-global-angular